🚚 Rename: 테스트 페이지들을 자동으로 관리하는 파일 이름 수정 및 이동 (#96)
