+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 // utils
 import { cn } from '@/utils/utils'
+import { testPages } from '@/utils/registryTestPages'
 
-export { cn }
+export { cn, testPages }
